Remove process protection code for maintainability and security reasons
Removed process protection code (protectprocess.c/h and related
function calls) from the project, and removed the DelayLoadDLL linking
from project settings as the associated DLL became unnecessary with the
removal of the protection code.

The code was originally intended to enhance security, but the
lack of an active maintainer and the presence of compiler warnings
raised concerns that it might inadvertently increase security risks.

Furthermore, this functionality is unrelated to core FTP
communication and may impede future refactoring efforts. It can be
reintroduced later if its benefits outweigh its risks.